+++ src/1.Framework/Vita/2.Operations/Contexts/WebCallContext.cs
@@ -55,7 +55,7 @@ namespace Vita.Entities.Api {
   /// <remarks>
   /// </remarks>
   public class WebCallContext {
-    public readonly Guid Id = Guid.NewGuid(); 
+    public readonly Guid Id = EntityApp.GuidFactory(); 
     // Used as key to save in Request properties. 
     public const string WebCallContextKey = "_vita_web_call_context_";
     public const string OperationContextKey = "_vita_operation_context_";