+++ src/1.Framework/Vita/2.Operations/SearchObjects.cs
@@ -6,8 +6,29 @@ using System.Threading.Tasks;
 using Vita.Entities.Api;
 
 namespace Vita.Entities {
-  /// <summary>Base class for search parameters containers. 
-  /// Use a derived class as a parameter for search methods in controllers.</summary>
+
+  /// <summary>Base interface for search parameters containers. Use an implementing type as a parameter for search methods in controllers.</summary>
+  /// <remarks>
+  /// <para>The <c>"IEntitySession.ExecuteSearch"</c> expects an instance of ISearchParams.  </para>
+  /// </remarks>
+  public interface ISearchParams {
+    /// <summary>List of columns with optional 'desc' spec. </summary>
+    /// <remarks>Use comma or semicolon to separate properties; use dash (-) or colon to add 'desc' flag (dash is more URL-friendly).
+    /// You can also map simple names to more complex names (or dotted property chains), by providing 
+    /// mapping dictionary to <c>"IEntitySession.ExecuteSearch"</c> method. </remarks>
+    /// <example>
+    ///   firstname-desc;lastname;birthdate-desc
+    /// </example>
+    string OrderBy { get; set; }
+    /// <summary>Number of rows to skip.</summary>
+    int Skip { get; set; }
+    /// <summary>Number of rows to take.</summary>
+    int Take { get; set; }
+  }
+
+
+
+  /// <summary>Default implmentation of <see cref="ISearchParams"/> interface. Use a derived class as a parameter for search methods.</summary>
   /// <remarks>
   /// <para>The <c>"IEntitySession.ExecuteSearch"</c> expects an instance of SearchParams sub-class.  </para>
   /// <para> Use the <see cref="EntitySessionExtensions.DefaultIfNull"/> extension method at the beginning 
@@ -15,7 +36,7 @@ namespace Vita.Entities {
   /// and/or to enforce (limit) maximum Take value. </para>
   /// <para>Important: Use properties in the derived class, not fields - fields do not work with Web Api [FromUri] attribute.</para>
   /// </remarks>
-  public class SearchParams {
+  public class SearchParams: ISearchParams {
     /// <summary>List of columns with optional 'desc' spec. </summary>
     /// <remarks>Use comma or semicolon to separate properties; use dash (-) or colon to add 'desc' flag (dash is more URL-friendly).
     /// You can also map simple names to more complex names (or dotted property chains), by providing