+++ src/1.Framework/Vita/4.Internals/Entities.Model/Construction/AttributesImplementation/EntityAttributesImpl_AutoAttr.cs
@@ -181,7 +181,7 @@ namespace Vita.Entities {
       var rec = record;
       if (rec.SuppressAutoValues)
         return;
-      var newGuid = Guid.NewGuid();
+      var newGuid = EntityApp.GuidFactory(); 
       rec.SetValueDirect(HostMember, newGuid);
     }